serve: Add "local" mode
Define an interface for the observer server, and add a new implementation
LocalObserverServer that simply ignores all the state change events. It's
meant to be started embeeded in Cilium process, and thus the parser can
access the Cilium state directly without maintaining a copy.

Note that LocalObserverServer still connects to the monitor socket to
receive monitor events from Cilium. We need to figure out how to access
monitor events from Cilium's ring buffer.

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>